web: Added ActivityPanel
